moved the functionality to server/rest-api
